+++ app/src/main/java/org/simple/clinic/home/overdue/OverdueEvent.kt
@@ -8,8 +8,6 @@ sealed class OverdueEvent : UiEvent
 
 data class CurrentFacilityLoaded(val facility: Facility) : OverdueEvent()
 
-data class OverdueAppointmentsLoaded(val appointments: List<OverdueAppointment>) : OverdueEvent()
-
 data class CallPatientClicked(val patientUuid: UUID) : OverdueEvent() {
   override val analyticsName = "Overdue Screen:Call Patient clicked"
 }