+++ app/src/main/java/org/simple/clinic/home/overdue/OverdueScreen.kt
@@ -1,11 +1,14 @@
 package org.simple.clinic.home.overdue
 
+import android.annotation.SuppressLint
 import android.content.Context
 import android.os.Parcelable
 import android.util.AttributeSet
 import android.widget.RelativeLayout
 import androidx.appcompat.app.AppCompatActivity
+import androidx.paging.PagedList
 import androidx.recyclerview.widget.LinearLayoutManager
+import com.jakewharton.rxbinding3.view.detaches
 import io.reactivex.rxkotlin.ofType
 import kotlinx.android.synthetic.main.screen_overdue.view.*
 import org.simple.clinic.ReportAnalyticsEvents
@@ -15,7 +18,7 @@ import org.simple.clinic.mobius.MobiusDelegate
 import org.simple.clinic.router.screen.ScreenRouter
 import org.simple.clinic.util.UserClock
 import org.simple.clinic.util.unsafeLazy
-import org.simple.clinic.widgets.ItemAdapter
+import org.simple.clinic.widgets.PagingItemAdapter
 import org.simple.clinic.widgets.visibleOrGone
 import java.time.LocalDate
 import java.time.format.DateTimeFormatter
@@ -26,7 +29,7 @@ import javax.inject.Named
 class OverdueScreen(
     context: Context,
     attrs: AttributeSet
-) : RelativeLayout(context, attrs), OverdueUi, OverdueUiActions {
+) : RelativeLayout(context, attrs), OverdueUiActions {
 
   @Inject
   lateinit var activity: AppCompatActivity
@@ -43,7 +46,10 @@ class OverdueScreen(
   @Inject
   lateinit var effectHandlerFactory: OverdueEffectHandler.Factory
 
-  private val overdueListAdapter = ItemAdapter(OverdueAppointmentRow.DiffCallback())
+  @field:[Inject Named("for_overdue_appointments")]
+  lateinit var pagedListConfig: PagedList.Config
+
+  private val overdueListAdapter = PagingItemAdapter(OverdueAppointmentRow.DiffCallback())
 
   private val events by unsafeLazy {
     overdueListAdapter
@@ -53,8 +59,6 @@ class OverdueScreen(
   }
 
   private val delegate by unsafeLazy {
-    val uiRenderer = OverdueUiRenderer(this)
-
     val date = LocalDate.now(userClock)
 
     MobiusDelegate.forView(
@@ -63,7 +67,7 @@ class OverdueScreen(
         update = OverdueUpdate(date),
         effectHandler = effectHandlerFactory.create(this).build(),
         init = OverdueInit(date),
-        modelUpdateListener = uiRenderer::render
+        modelUpdateListener = { /* Nothing to do here */ }
     )
   }
 
@@ -97,21 +101,24 @@ class OverdueScreen(
     super.onRestoreInstanceState(delegate.onRestoreInstanceState(state))
   }
 
-  override fun updateList(overdueAppointments: List<OverdueAppointment>, isDiabetesManagementEnabled: Boolean) {
-    val areOverdueAppointmentsAvailable = overdueAppointments.isNotEmpty()
-    viewForEmptyList.visibleOrGone(isVisible = !areOverdueAppointmentsAvailable)
-    overdueRecyclerView.visibleOrGone(isVisible = areOverdueAppointmentsAvailable)
-
-    overdueListAdapter.submitList(OverdueAppointmentRow.from(
-        appointments = overdueAppointments,
-        clock = userClock,
-        dateFormatter = dateFormatter,
-        isDiabetesManagementEnabled = isDiabetesManagementEnabled
-    ))
-  }
-
   override fun openPhoneMaskBottomSheet(patientUuid: UUID) {
     activity.startActivity(ContactPatientBottomSheet.intent(context, patientUuid))
+  }
+
+  @SuppressLint("CheckResult")
+  override fun showOverdueAppointments(dataSource: OverdueAppointmentRowDataSource.Factory) {
+    val detaches = detaches()
+
+    dataSource
+        .toObservable(pagedListConfig, detaches)
+        .takeUntil(detaches)
+        .doOnNext { appointmentsList ->
+          val areOverdueAppointmentsAvailable = appointmentsList.isNotEmpty()
+
+          viewForEmptyList.visibleOrGone(isVisible = !areOverdueAppointmentsAvailable)
+          overdueRecyclerView.visibleOrGone(isVisible = areOverdueAppointmentsAvailable)
+        }
+        .subscribe(overdueListAdapter::submitList)
   }
 
   interface Injector {