+++ app/src/main/java/org/simple/clinic/overdue/AppointmentRepository.kt
@@ -1,5 +1,6 @@
 package org.simple.clinic.overdue
 
+import androidx.paging.PositionalDataSource
 import io.reactivex.Completable
 import io.reactivex.Observable
 import io.reactivex.Single
@@ -127,6 +128,19 @@ class AppointmentRepository @Inject constructor(
         scheduledBefore = since,
         scheduledAfter = since.minus(appointmentConfig.periodForIncludingOverdueAppointments)
     )
+  }
+
+  fun overdueAppointmentsDataSource(
+      since: LocalDate,
+      facility: Facility
+  ): PositionalDataSource<OverdueAppointment> {
+    return overdueDao
+        .overdueAtFacilityDataSource(
+            facilityUuid = facility.uuid,
+            scheduledBefore = since,
+            scheduledAfter = since.minus(appointmentConfig.periodForIncludingOverdueAppointments)
+        )
+        .create() as PositionalDataSource<OverdueAppointment>
   }
 
   fun overdueAppointmentsCount(